ARCO-204: fix PR comments
